Add a copy of StatusOr and associated tests
Building on publicly available prior art, including:
* https://github.com/protocolbuffers/protobuf/blob/master/src/google/protobuf/stubs/statusor.h
* https://github.com/tensorflow/tensorflow/blob/master/tensorflow/stream_executor/lib/statusor.h

We write our own as opposed to leveraging any of these because each of
these projects also uses a project-specific Status type that predates
the public availability of absl::Status.

Most of this change will be obsolete and can be deleted when Abseil
releases a canonical StatusOr publicly. This is already in the works
(see go/absl-statusor).

Change-Id: Ie506f2dba620f6c269ae3308d88482b3a88e7000