Generate bridge.h and main.cc from pkcs11f.h
There are a bunch of places where we'll need to end up doing something
"for all of the Cryptoki functions":
* Writing the bridge header, and the main impl (this commit)
* Pulling function pointers into a CK_FUNCTION_LIST
* Writing platform-specific stuff for exporting symbols on Linux, Mac,
and Windows

Parsing the function declarations out of the C headers so that we can
generate all of those seems like the least crazy solution for keeping
everything in sync. Unfortunately, this can't be done just with macros
because the function arguments are not part of the macro invocations.

As well, we'd have to do it all again -- and possibly manage two
versions -- if we ever want to support the 3.0 Cryptoki API.

Bug: b/148377771
Change-Id: I5cf586ce253e875ed14c63d72cd4b96d620f3b3e