Add support for converting YAML to protobuf
We need a configuration language, and sadly, textproto doesn't fit the
bill. The format isn't well documented publicly, and also, the public
textproto parser doesn't emit parsing errors on malformed input-- it
just quietly fails, and lets you guess what might be wrong. :-(

But ultimately we still need a schema and an object model, and so we
might as well leverage protobuf fro that. This little bridge essentially
uses protobuf message descriptors as a YAML schema, and emits reasonable
parsing errors when it fails.

Doing something similar with structs and normal yaml-cpp constructs
ultimately ends up being more code-- and ugly code too, because of
yaml-cpp's heavy use of exceptions in non-exceptional cases.

Change-Id: Ia201bfaf945a4c43830eb20021c1eeafae41988d