Upgrade to github.com/golang/protobuf >= v1.40
v1.40 is the version that contains the v2 bridge that allows proto
reflection.

Practically, this is accomplished by upgrading the Bazel rules for Go,
which include protobuf as a dependency. (It is also possible to override
the protobuf library explicitly, as documented at [0], but the latest
rules-go release contains appropriate support.)

As part of the upgrade, some existing code needs to be tweaked to use
the new protocmp routines, because the protobuf upgrade introduces
unexported struct fields.

[0] https://github.com/bazelbuild/rules_go/blob/master/go/dependencies.rst#overriding-dependencies

Change-Id: Icd98a96af4b57f600a7f4f035ede6eb862b42d90