Support forking and module loading/unloading
Back story: gRPC creates background threads that do things. If those
background threads are working when our library is unloaded, they will
cause the process to segfault. If those background threads hold internal
mutexes when our library is forked, then those mutexes must be cleared
before gRPC is functional in the child process.

Module loading/unloading should be understood as a baseline requirement
for the alpha, and forking is a nice-to-have that we'll want later. But
since they touch the same code points, let's do it all now.

The actual work here is to ensure that gRPC is shut down when our
library is shut down, and that gRPC is re-initialized when our library
is re-initialized after a fork.

b/160312835 tracks adding tests for this post-alpha.

Change-Id: I968a1a878ff303928fc27b5bb1cc16dd71706d23