Separate certificate generation out of certificate object creation
In order to support re-scrape, we'll update the logic that builds a view
of KMS state to return a proto as an intermediate output. We'll use that
proto to build our Cryptoki objects. This makes it straightforward to
test the functionality that builds a view of KMS' state.

To support this approach, any library state that ought to remain
unchanged when KMS state changes will need to be moved into that layer.
This includes generating object handles, and generating certificates.

Bug: b/161536656
Change-Id: Ibaa46cd9bdde78a65b98573f61418882f75ddf11