Add ObjectStoreState proto to represent a snapshot of library state
This provides a clear separation between "the thing that translates KMS
state into PKCS #11 state" and "the thing that creates PKCS #11 objects
from a defined PKCS #11 state".

This is a necessary step for refreshing based on updated state in KMS,
because we need to ensure that object handles remain unchanged when
state is updated.

Bug: b/161536656
Change-Id: I1639832ceef0fef32d3270f7d30a63d6f3b287c1