Add tests for the binaries we produce
Add a change detector test for binary size on all platforms.

On Linux and FreeBSD, test to ensure that we're not exporting or
importing symbols that we don't intend to.

As part of this change, eliminate our custom toolchain for Linux. It
turns out we can get our binary the way we want it without a custom
toolchain.

Bug: b/160314006
Change-Id: If361ce77dc04368a2cbb35ee80a2305143358b41