Add build signing for dev builds
Implemented as ECDSA signing with a KMS key, as discussed in
b/184264411.

Just doing this with gcloud is more trouble than it's worth, because our
build machines have some ancient gcloud installed, and updating it takes
forver.

Still TODO: build signing for release builds, with a key that is more
locked down.

Bug: b/160886831
Change-Id: I4f365450f41ac5b7ded111f8785476df8f817feb