Rename generate_certs to experimental_generate_certs
We added this feature based on a now obsolete request; I believe it may
be totally unused. I'd like to call it "experimental" in the GA so that
we have some leeway to drop it entirely if it has no users.

Also, add documentation for log_filename_suffix, which was mistakenly
undocumented.

Change-Id: Ib1fc3cc04a9ad629f236ae368b110fea954fa598