Add key caching stress test.
The KMS state is refreshed every second, with a key being created and
destroyed approximately within the same interval, using fakekms.

This initial draft of the test chooses a valid KMS algorithm at random,
and creates a single version of every key. Creating a single key with
multiple versions is also an option, and can be added either as a
replacement or as an additional choice in a follow-up CL (the config
would then need to include experimental_create_multiple_versions).

BUG: b/175422038
Change-Id: Id6748147a94c67ab11ef8f42242c757644389416