Correct the returned value for CKA_EC_POINT.
The definition of CKA_EC_POINT calls out to ANSI X9.62, which specifies
that the serialized point value must be encoded in an ASN.1 octet string.

Confusingly, X9.62 calls the plain serialization an "octet string", but
it's not ASN.1. Identified by a user at
https://github.com/GoogleCloudPlatform/kms-integrations/issues/1.

It turns out we are not the only people to make this mistake. The JDK's
initial implementation had this bug [0], and now accepts either form for
compatibility purposes [1]. NSS also accepts either form, for compatibility
purposes [2].

[0] https://bugs.java.com/bugdatabase/view_bug.do?bug_id=6763530
[1] https://github.com/openjdk/jdk/blob/7326481143c1321700cbf2caa9e068c5077e22c4/src/jdk.crypto.cryptoki/share/classes/sun/security/pkcs11/P11Key.java#L1098
[2] https://bugzilla.mozilla.org/show_bug.cgi?id=480280

Change-Id: Icd9cf73a0b06c10a3557a1e174ab72b500d6b83b