Patch interoperable AES into dependency stack.
This change patches both resources (the algorithm enum) and service (the
actual RPCs). To do the latter, the google_cloud_go KMS client also
needs to be patched using the same `git diff --patch` technique. The
diffs have to be applied using `-p2` because of how the repo is
structured, and the patch needs to be part of the go_repository rule
that's fetching the repo from cloud.google.com (/not/ Github), to avoid
running into issues with gazelle regarding missing BUILD files.

I have also updated a couple dependencies to more recent versions (eg.
rules_go, googleapis, ..).

Bug: b/234842124
Change-Id: I81efbe82d7476f4b23ae42af1f11c4866358efd5