Refactor CK_GCM_PARAMS parsing logic.
Perform struct conversion inside the validation helper to go through
that logic only once. If the user is using the errata version of the
struct, convert it to the pre-errata version by fixing the fields and
adding the ulIvBits value manually based on ulIvLen.

Move errata struct definition to cryptoki.h to avoid struct packing
issues on Windows.

Bug: b/234842124
Change-Id: I6582101c1e0adaff7fdeb11c607b63f836d190c9