Add support for multi-part AES-GCM encryption/decryption.
Adding support for both at the same time to be able to test the new
logic. Skip the part where C_EncryptUpdate should be returning the
encrypted data part since AES-GCM uses AEAD and also the fact that we
would be generating a different IV for every single encrypted block.

RELEASE_NOTES

Bug: b/234842124
Change-Id: I594feadaa5252fccbc4d4a7395b220cdf922f09c