Add checksum library based on new crc32c dependency.
This change involves adding a dependency on the opensource Google crc32c
library, which requires custom BUILD files since the library was
designed to be built with cmake, not bazel. I'm reusing the BUILD files
by google-cloud-cpp, IIRC we removed that dependency a while ago so it's
porbably not worth adding it back just for this (let me know if you
disagree).

You can find the original files here:
https://github.com/googleapis/google-cloud-cpp/tree/main/bazel

Bug: b/160311385
Change-Id: I12442988845401f9de8890379f56981a80ffe3ab