Fix C_Initialize init_args evaluation.
Per the general-purpose function section of the spec, C_Initialize
specifies 4 different possible flags+params combinations that represent
different multi-threading expectations by user applications.

In the current implementation, the scenario where init_args->flags == 0
and the function pointers are null is rejected as invalid, despite being
a valid configuration supported by our library.

Case #1 in the spec says this configuration "means that the application
won’t be accessing the Cryptoki library from multiple threads
simultaneously", which seems okay. Our library always creates multiple
threads because we are using gRPC, but as long as the application isn't
passing CKF_LIBRARY_CANT_CREATE_OS_THREADS (at which point we rightfully
return an error), this behavior is spec-compliant.

The tl;dr is that we will only error in the case (3) from the relevant
section of the spec linked below.

https://docs.oasis-open.org/pkcs11/pkcs11-base/v2.40/os/pkcs11-base-v2.40-os.html#_Toc385057894

RELEASE_NOTES

Change-Id: I2e42673a06020e46e9ef4f1b3a99748ec807c465