Migrate from google/crc32c dep to Abseil's crc32c.
This gets rid of a bunch of custom build files and simplifies integrity
verification in the library. I'm keeping the dedicated util library for
convenience.

Note: the latest Abseil release is only supported on C++14 or higher,
and requires (for some unclear reasons) the host_* flags to correctly
set the C++ version to use at build time.

Bug: b/160311385
Change-Id: I84ce6bd8e828bd5ee09b7d912aa98819dffe319b