Add helper to register/unregister provider and related tests.
The provider registration helper is based on the POC put together by
bdhess@, tweaked to register/unregister the provider automatically.

To register the provider correctly, the DLL needs to be in System32,
so we copy it to System32 using robocopy in a custom genrule with a
batch command.

NOTE: the batch command explicitly returns 0, since bazel expects 0 but
robocopy returns 1 when "one or more files were copied successfully",
causing the build to fail.

Also add check for provider name to bridge.

Bug: b/270418439
Change-Id: I559c0a51254d4ac37018517b74d4e8467d1ea92f