Add support for NCryptEnumKeys.
Note: I have added an optional parameter to EnumKeys, which lets us
specify a custom config file path instead of using the hardcoded one.
This is because I wanted to avoid messing with existing config files in
that path, and also to avoid having yet another env variable to rely on.
But I guess that's an alternative we can also consider, which is easy to
leverage in our tests too.

Bug: b/287672102
Change-Id: I2e10e59203e80ec36d1bb16dffa542cb03c14c3f