Drop Visual C++ installation check and update user guide.
I took a closer look at this and found a helpful doc page:
https://learn.microsoft.com/en-US/cpp/windows/latest-supported-vc-redist?view=msvc-170#notes.

The Visual C++ version should be the same (or higher) as the version
used to build the provider, which in our case is 14.34 because IIUC
that's what comes pre-installed in the windows2019 kokoro image.

Let's drop the check for now as discussed, which results in the
not-as-friendly error message that I have added to the user guide.
At least it won't block user installation.

Note that the installer says "2015-2022" because
"Visual Studio versions since Visual Studio 2015 share the same
Redistributable files", per the link I shared above.

Also updated the user guide doc to reflect recent changes, and fixed the new
installer tests.

Bug: b/288279385
Bug: b/280633307
Change-Id: Ice8ee6c5e6828bcb66b8410d36f22502db59b57b