Drop experimental raw AES and HMAC config flags.
For a future v1.3 release, we won't need the raw AES and HMAC flags
anymore. We could also rename the raw AES flag and keep it around as a
way to opt out of the expert feature, but that's already covered by the
IAM permission so I don't see much value there.

RELEASE_NOTES

Bug: b/285611641
Change-Id: I589460ff5f58f9c5b3b4185cc4b6bc5323199e9f