Bump CNG provider version to 0.9
The recent system roots patch seems like a good enough improvement to
bump the version already, and I'm hoping we'll soon be able to fix the
ADCS integration too. Also, I'd like to add a separate provider version
to the cloudkms-cep-lab bucket and test it with the upcoming CNG prober.

Change-Id: I61e22aea552cf7f986df73a3224f0130defe05e6