Fix potential undefined behavior in FindOp::Next.
When there are no more items left in the objects_ vector, we currently
still return a zero-size Span that uses the address of the out-of-bounds
value of offset_ (previously set to objects_.size()).

This may cause UB errors and issues such as
https://github.com/GoogleCloudPlatform/kms-integrations/issues/50, which
is possibly due to the library being built with additional hardening
options (we don't support Fedora as a target platform so there are no
available pre-compiled builds).

Explictly return an empty, default-constructed Span instead.

Bug: b/425872666
Change-Id: I88a521e231f553412d785dfa0dc78d63aa978be8