Fix return value handling in Windows build script.
e2e_test errors were being silently ignored in previous builds! I'm
still not 100% sure why, but early returning seems to fix this, and it
doesn't really make sense to keep going with the build if something
fails, as long as we make sure to upload testlogs first.

* Sample silent failure: http://fusion2/ci/kokoro/prod%3Acloud-kms%2Foss-tools%2Fpresubmit%2Fwindows_amd64/activity/eb7f1505-2932-45c1-a33b-017bdfe977c5
* Intentional test error now being detected: http://fusion2/ci/kokoro/prod%3Acloud-kms%2Foss-tools%2Fpresubmit%2Fwindows_amd64/activity/b411ff04-a0e5-4ad9-94d3-29c9c03b33cc

Note: this will increase the number of flakes we see, because e2e_test
can flake in similar ways to the signtool and install tests.

Change-Id: I36af0ecaba448bf4eb75fe481f353decc351f620